Make ObjcAccess an object safe trait, hopefully
Note the api has been reworked.
The get_from_backing_obj() method has been changed to get_backing_obj
and now returns a Ref to the Object rather then calling a callback
with it.
This is because there was no way to make the callback method work
without generics but we can't have generics if we want to make it
object safe